+++ Kernel/System/Loader.pm
@@ -1,5 +1,6 @@
 # --
 # Copyright (C) 2001-2021 OTRS AG, https://otrs.com/
+# Copyright (C) 2021 Znuny GmbH, https://znuny.org/
 # --
 # This software comes with ABSOLUTELY NO WARRANTY. For details, see
 # the enclosed file COPYING for license information (GPL). If you
@@ -11,7 +12,7 @@ package Kernel::System::Loader;
 use strict;
 use warnings;
 
-use CSS::Minifier qw();
+use CSS::Minifier qw();    # default minifier, will only be used if CSS::Minifier:XS is not available.
 use JavaScript::Minifier qw();
 
 our @ObjectDependencies = (
@@ -340,7 +341,15 @@ sub MinifyCSS {
         return;
     }
 
-    my $Result = CSS::Minifier::minify( input => $Param{Code} );
+    my $Result;
+
+    my $IsCSSMinifierXSAvailable = $Self->IsCSSMinifierXSAvailable();
+    if ($IsCSSMinifierXSAvailable) {
+        $Result = CSS::Minifier::XS::minify( $Param{Code} );
+    }
+    else {
+        $Result = CSS::Minifier::minify( input => $Param{Code} );
+    }
 
     # a few optimizations can be made for the minified CSS that CSS::Minifier doesn't yet do
 
@@ -520,6 +529,30 @@ sub CacheDelete {
     );
 
     return @Result;
+}
+
+=head2 IsCSSMinifierXSAvailable()
+
+    Tries to load CSS::Minifier::XS if available which provides faster creation of minified CSS.
+
+    Returns true value if CSS::Minifier::XS is available and loaded.
+
+=cut
+
+sub IsCSSMinifierXSAvailable {
+    my ( $Self, $Param ) = @_;
+
+    return $Self->{CSSMinifierXSAvailable} if defined $Self->{CSSMinifierXSAvailable};
+
+    $Self->{CSSMinifierXSAvailable} = eval {
+        require CSS::Minifier::XS;
+        CSS::Minifier::XS->import();
+        1;
+    };
+
+    $Self->{CSSMinifierXSAvailable} //= 0;
+
+    return $Self->{CSSMinifierXSAvailable};
 }
 
 1;