+++ scripts/test/Loader.t
@@ -1,5 +1,6 @@
 # --
 # Copyright (C) 2001-2021 OTRS AG, https://otrs.com/
+# Copyright (C) 2021 Znuny GmbH, https://znuny.org/
 # --
 # This software comes with ABSOLUTELY NO WARRANTY. For details, see
 # the enclosed file COPYING for license information (GPL). If you
@@ -34,8 +35,13 @@ my $Home = $ConfigObject->Get('Home');
 
     $CSS = ${$CSS};
 
+    my $ExpectedMinifiedFilename = 'OTRS.Reset.min_for_css_minifier.css';
+    if ( $LoaderObject->IsCSSMinifierXSAvailable() ) {
+        $ExpectedMinifiedFilename = 'OTRS.Reset.min_for_css_minifier_xs.css';
+    }
+
     my $ExpectedCSS = $MainObject->FileRead(
-        Location => $Home . '/scripts/test/sample/Loader/OTRS.Reset.min.css',
+        Location => "$Home/scripts/test/sample/Loader/$ExpectedMinifiedFilename",
     );
 
     $ExpectedCSS = ${$ExpectedCSS};